requirements.txt
# ========================================
# 核心深度学习框架
# ========================================
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# ========================================
# Transformers生态
# ========================================
transformers>=4.36.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
peft>=0.7.0
safetensors>=0.4.0

# ========================================
# 数据处理
# ========================================
datasets>=2.15.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# ========================================
# 可视化
# ========================================
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0

# ========================================
# 机器学习工具
# ========================================
scikit-learn>=1.3.0

# ========================================
# 实验管理
# ========================================
wandb>=0.16.0
tensorboard>=2.15.0

# ========================================
# 工具库
# ========================================
tqdm>=4.66.0
pyyaml>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0
rich>=13.0.0

# ========================================
# Jupyter支持
# ========================================
jupyter>=1.0.0
ipywidgets>=8.1.0
notebook>=7.0.0

# ========================================
# 开发工具
# ========================================
pytest>=7.4.0
black>=23.12.0
isort>=5.13.0

# ========================================
# API和服务
# ========================================
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
huggingface_hub>=0.19.0
